+++ app/src/main/assets/plant_descriptions/hance.txt
@@ -1,0 +1,1 @@
+﻿这是黄馨
