+++ app/src/main/assets/plant_descriptions/jacq.txt
@@ -1,0 +1,1 @@
+﻿这是月季
