+++ app/src/main/assets/plant_descriptions/linn.txt
@@ -1,0 +1,1 @@
+﻿这是广玉兰
