+++ app/src/main/assets/plant_descriptions/photinia.txt
@@ -1,0 +1,1 @@
+﻿这是石楠花
