+++ app/src/main/assets/plant_descriptions/planch.txt
@@ -1,0 +1,1 @@
+﻿这是杜鹃
