--- app/src/main/assets/plant_descriptions/plant1.txt
@@ -1,1 +1,0 @@
-植物
