--- app/src/main/assets/plant_descriptions/plant2.txt
@@ -1,1 +1,0 @@
-1267361
