+++ app/src/main/java/com/example/tony/finalproject_plants/moreclass.java
@@ -1,0 +1,10 @@
+package finalproject_plants;
+
+import android.app.Fragment;
+
+/**
+ * Created by user on 2017/10/23.
+ */
+
+public class moreclass extends Fragment {
+}
