+++ app/src/main/java/com/example/tony/finalproject_plants/noteclass.java
