--- app/src/main/java/finalproject_plants/findclass.java
@@ -1,10 +1,0 @@
-package finalproject_plants;
-
-import android.app.Fragment;
-
-/**
- * Created by user on 2017/10/23.
- */
-
-public class findclass extends Fragment {
-}
